[#126] Rename subscribe classes to purchase.
