+++ apigee_m10n.routing.yml
@@ -141,8 +141,8 @@ entity.rate_plan.canonical:
 entity.rate_plan.subscribe:
   path: /user/{user}/monetization/package/{package}/plan/{rate_plan}/purchase
   defaults:
-    _controller: \Drupal\apigee_m10n\Entity\Controller\RatePlanSubscribeController::subscribeForm
-    _title_callback: \Drupal\apigee_m10n\Entity\Controller\RatePlanSubscribeController::title
+    _controller: \Drupal\apigee_m10n\Entity\Controller\PurchaseRatePlanController::purchaseForm
+    _title_callback: \Drupal\apigee_m10n\Entity\Controller\PurchaseRatePlanController::title
   requirements:
     user: '^[1-9]+[0-9]*$'
     _entity_developer_access: rate_plan.subscribe
