+++ modules/apigee_m10n_teams/src/Entity/Controller/TeamPurchaseRatePlanController.php
@@ -17,55 +17,25 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-namespace Drupal\apigee_m10n\Entity\Controller;
+namespace Drupal\apigee_m10n_teams\Entity\Controller;
 
-use Apigee\Edge\Api\Monetization\Entity\Developer;
+use Apigee\Edge\Api\Monetization\Entity\Company;
+use Drupal\apigee_edge_teams\Entity\TeamInterface;
+use Drupal\apigee_m10n\Entity\Controller\PurchaseRatePlanController;
 use Drupal\apigee_m10n\Entity\RatePlanInterface;
 use Drupal\apigee_m10n\Entity\PurchasedPlan;
 use Drupal\apigee_m10n\Form\PurchasedPlanConfigForm;
-use Drupal\Core\Controller\ControllerBase;
-use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
-use Drupal\Core\Entity\EntityFormBuilderInterface;
-use Drupal\Core\Routing\RouteMatchInterface;
-use Drupal\user\UserInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Controller for subscribing to rate plans.
  */
-class RatePlanSubscribeController extends ControllerBase implements ContainerInjectionInterface {
+class TeamPurchaseRatePlanController extends PurchaseRatePlanController {
 
   /**
-   * Entity form builder.
+   * Page callback to create a new team purchased_plan.
    *
-   * @var \Drupal\Core\Entity\EntityFormBuilderInterface
-   */
-  protected $entityFormBuilder;
-
-  /**
-   * BillingController constructor.
-   *
-   * @param \Drupal\Core\Entity\EntityFormBuilderInterface $entityFormBuilder
-   *   Entity form builder service.
-   */
-  public function __construct(EntityFormBuilderInterface $entityFormBuilder) {
-    $this->entityFormBuilder = $entityFormBuilder;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('entity.form_builder')
-    );
-  }
-
-  /**
-   * Page callback to create a new purchased_plan.
-   *
-   * @param \Drupal\user\UserInterface $user
-   *   The user that will be purchasing the plan.
+   * @param \Drupal\apigee_edge_teams\Entity\TeamInterface $team
+   *   The team purchasing the rate plan.
    * @param \Drupal\apigee_m10n\Entity\RatePlanInterface $rate_plan
    *   The rate plan.
    *
@@ -74,23 +44,24 @@ class RatePlanSubscribeController extends ControllerBase implements ContainerInj
    *
    * @throws \Exception
    */
-  public function subscribeForm(UserInterface $user, RatePlanInterface $rate_plan) {
+  public function teamPurchaseForm(TeamInterface $team, RatePlanInterface $rate_plan) {
     // Create a purchased_plan to pass to the purchased_plan edit form.
     $purchased_plan = PurchasedPlan::create([
       'ratePlan' => $rate_plan,
-      'developer' => new Developer(['email' => $user->getEmail()]),
+      'company' => new Company(['id' => $team->id()]),
       'startDate' => new \DateTimeImmutable(),
     ]);
 
     // Get the save label from settings.
     $save_label = $this->config(PurchasedPlanConfigForm::CONFIG_NAME)->get('purchase_button_label');
-    $save_label = $save_label ?? 'Subscribe';
+    $save_label = $save_label ?? 'Purchase';
 
     // Return the purchase form with the label set.
     return $this->entityFormBuilder->getForm($purchased_plan, 'default', [
       'save_label' => $this->t($save_label, [
         '@rate_plan' => $rate_plan->getDisplayName(),
-        '@username' => $user->label(),
+        '@team' => $team->label(),
+        '@username' => '',
       ]),
     ]);
   }
@@ -98,24 +69,25 @@ class RatePlanSubscribeController extends ControllerBase implements ContainerInj
   /**
    * Gets the title for the purchase page.
    *
-   * @param \Drupal\Core\Routing\RouteMatchInterface $route_match
-   *   The route match.
-   * @param \Drupal\user\UserInterface|null $user
-   *   The user.
-   * @param \Drupal\apigee_m10n\Entity\RatePlanInterface|null $rate_plan
+   * @param \Drupal\apigee_edge_teams\Entity\TeamInterface $team
+   *   The team purchasing the rate plan.
+   * @param \Drupal\apigee_m10n\Entity\RatePlanInterface $rate_plan
    *   The rate plan.
    *
    * @return \Drupal\Core\StringTranslation\TranslatableMarkup
    *   The title.
    */
-  public function title(RouteMatchInterface $route_match, UserInterface $user = NULL, RatePlanInterface $rate_plan = NULL) {
+  public function teamTitle(TeamInterface $team, RatePlanInterface $rate_plan) {
     $title_template = $this->config(PurchasedPlanConfigForm::CONFIG_NAME)->get('purchase_form_title');
     $title_template = $title_template ?? 'Purchase @rate_plan';
+    // TODO: Add information about the availability of `@teamname`.
     return $this->t($title_template, [
       '@rate_plan' => $rate_plan->getDisplayName(),
       '%rate_plan' => $rate_plan->getDisplayName(),
-      '@username' => $user->label(),
-      '%username' => $user->label(),
+      '@teamname' => $team->label(),
+      '%teamname' => $team->label(),
+      '@username' => '',
+      '%username' => '',
     ]);
   }
